cga: Do not write a character when moving the cursor
A naive implementation to solve the white foreground problem would write
a white foreground everytime the cursor moves. However, there are some
characters like '\b' that only move the cursor without overwriting
anything.

Change the handling for color to be done instead when initializing
memory and scrolling, so that any new cells have a white foreground,
ensuring we can always see the cursor.