Add test directory for our kernel tests
We can copy most tests from the `user/tests` directory. These tests are
supposed to be run in usermode once we finish phase 5. However, we'd
still like to test our code, so this is a temporary solution until we
get it working:

- test directly in start.c.
- create a bunch of test processes in test-kernel.

Once phase 5 is done, we'll be able to use the tests in
`user/tests`.