task: Zombie processes
In free_zombie_tasks, make sure to free the stack. This leads to a
problem where we cannot free the stack before context switching, as the
context switch still requires the old stack to be allocated.

To work around this, we don't free a zombie process if it's about to
context switch.