Children are created
In the start method, during the creation of the task, we create a child.
We assume that the father of the created task is the current task.
The children of a task are store on the children attribute.
The siblings of the children are stored in the siblings attribute (we use it for the children queue)