shm: Initial implementation
Because we haven't implemented virtual memory, we can sidestep
a lot of the difficulty here.

This module is implemented using a hash
table and a refcount system. Processes that want to get a shared memory
table use shm_create(), registering the page with a key. References
can be acquired via shm_acquire(), and released via shm_release().

When a shared memory page has no more refs, it's freed automatically,
implemented with a simple refcounting system.