task: add set_task_ready_or_running(), start_test()
Separate set_task_ready into two functions. Users can call
set_task_ready_or_running() when they want their process to become
running if their process' prio is greater than the current running
processes prio.

Also fixes a really wierd bug with one of the tests. When using this
function in start(), test4() has a usermode program that runs in a loop.
For it to be interrupted, it uses a test_it() function that puts an
interrupt.

The problem however is located way back at the start: when we start the
test process itself (the one that runs test4_main), if, and ONLY IF we
schedule in that start() (like by running the new
set_task_ready_or_running) then the interrupt flag from eflags never
gets set back by this function, and the loop keeps running infinitely.

I tried to debug this for way too long and unfortunately, I figured out
the fix, but not why this happens in the first place.

To combat this, I added a start_test() function, that should be used
ONLY to run usermode tests. The purpose of this function is to run
test4 correctly, and any other tests that will use test_it(). start_test
() runs set_task_ready instead of set_task_ready_or_running, which means
 it will NOT context switch to the newly created process if it has an
 higher prio, which fixes the bug.

And for why this happens, I tried to understand the code in the
test_it() function, but I didn't understand at all unfortunately. It
puts EFLAGS (the register that holds the interrupt flag) on the
stack, and somehow it doens't work if it's 0x200?