task: Use a temporary kernel stack for __exit
When __exit is called, it will call some functions and assign some
variables, which will use the stack. The stack it will use is the stack
assigned to that process. However, if we have a very small stack size
assigned, for example if a process tries to start() with an empty stack
size, then __exit will try to use that stack and will crash.

Instead, use a global stack for every __exit invocation. However, we
can't switch the stack pointer to that global stack in the __exit
function directly, since the prologue may interfere by setting up the
stack (the prologue is the bit of assembly code run before every functon
 that GCC inserts).

Instead, we created our own assembly wrapper __exit, and renamed the
current function to __exit1. __exit's role is to move the stack
pointer, adding an offset (since the top of the stack is at the bottom
of the adresses) and then call __exit1.

This does not fix test6: we still have a problem with waitpid.