task: Check in __set_task_state if the task is not interrupted
Since msg.c manages its own queues, __set_task_state assumes the task
was in a queue. However, because of set_task_interrupted_msg
which does not add to a queue anymore, since the queue is managed by
msg.c now, the task is not in a queue, and calling queue_del crashes.

Check if it's a msg to avoid deleting.