schedule: do not switch to another task of lower priority
This is specified in https://ensiwiki.ensimag.fr/index
.php?title=Projet_syst%C3%A8me_:_Aspects_techniques
Another task should not run as long as there is an higher prio one.

So if we have a task with prio 2 running, and one with prio 1 waiting,
continue executing the prio 2 one.

This does not apply when the task is scheduling out, as at that point it
 should switch to any available process.