exit: order processes by their exit time when killed
for collection by waitpid(). When getting a bunch of processes that had
the same pid (for example rapid fire launch and return a value, so that
two or more processes have the same pid), then waitpid expects to get
each return value in the order the procesess died.

To manage this case, introduce an ordering on zombie processes, based on
 when they died. The process that died the earliest will be the first to
be returned.